Fix: Extract logic for getting sessions in App component.
