+++ src/api/postSessions.js
@@ -3,11 +3,9 @@ import axios from 'axios';
 const postSessions = async (username, email, pass, handleSuccesfullAuth) => {
   axios.post('http://localhost:3001/sessions',
     {
-      user: {
-        username,
-        email,
-        password: pass,
-      },
+      username,
+      email,
+      password: pass,
     },
     { withCredentials: true }).then(r => {
     if (r.data.logged_in) {
